services/core/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pydantic-settings>=2.0.0
email-validator>=2.0.0
redis>=5.0.0
httpx>=0.27.0
python-multipart>=0.0.9
